@@ -1,4 +1,4 @@
-// swift-tools-version:5.4
+// swift-tools-version:5.5.0
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -15,7 +15,8 @@ let excludeFiles: String = []
 let package = Package(
     name: "Web3swift",
     platforms: [
-        .macOS(.v10_12), .iOS(.v11)
+        .macOS(.v10_15),
+            .iOS(.v15)
     ],
     products: [
         .library(name: "web3swift", targets: ["web3swift"])
@@ -23,7 +24,6 @@ let package = Package(
     
     dependencies: [
         .package(url: "https://github.com/attaswift/BigInt.git", from: "5.3.0"),
-        .package(url: "https://github.com/mxcl/PromiseKit.git", from: "6.16.2"),
         .package(url: "https://github.com/daltoniam/Starscream.git", from: "4.0.4"),
         .package(url: "https://github.com/krzyzanowskim/CryptoSwift.git", from: "1.4.2")
     ],
@@ -31,7 +31,7 @@ let package = Package(
         .target(name: "secp256k1"),
         .target(
             name: "web3swift",
-            dependencies: ["BigInt", "secp256k1", "PromiseKit", "Starscream", "CryptoSwift"],
+            dependencies: ["BigInt", "secp256k1", "Starscream", "CryptoSwift"],
             exclude: excludeFiles,
             resources: [
                 .copy("./Browser/browser.js"),
