@@ -6,41 +6,39 @@
 
 import Foundation
 import BigInt
-//import EthereumAddress
-import PromiseKit
 
 //Token Standard
 protocol IERC20 {
-    func getBalance(account: EthereumAddress) throws -> BigUInt
-    func getAllowance(originalOwner: EthereumAddress, delegate: EthereumAddress) throws -> BigUInt
-    func transfer(from: EthereumAddress, to: EthereumAddress, amount: String) throws -> WriteTransaction
-    func transferFrom(from: EthereumAddress, to: EthereumAddress, originalOwner: EthereumAddress, amount: String) throws -> WriteTransaction
-    func setAllowance(from: EthereumAddress, to: EthereumAddress, newAmount: String) throws -> WriteTransaction
-    func approve(from: EthereumAddress, spender: EthereumAddress, amount: String) throws -> WriteTransaction
-    func totalSupply() throws -> BigUInt
+    func getBalance(account: EthereumAddress) async throws -> BigUInt
+    func getAllowance(originalOwner: EthereumAddress, delegate: EthereumAddress) async throws -> BigUInt
+    func transfer(from: EthereumAddress, to: EthereumAddress, amount: String) async throws -> WriteTransaction
+    func transferFrom(from: EthereumAddress, to: EthereumAddress, originalOwner: EthereumAddress, amount: String) async throws -> WriteTransaction
+    func setAllowance(from: EthereumAddress, to: EthereumAddress, newAmount: String) async throws -> WriteTransaction
+    func approve(from: EthereumAddress, spender: EthereumAddress, amount: String) async throws -> WriteTransaction
+    func totalSupply() async throws -> BigUInt
 }
 
 // This namespace contains functions to work with ERC20 tokens.
 // variables are lazyly evaluated or global token information (name, ticker, total supply)
 // can be imperatively read and saved
-public class ERC20: IERC20 {
-    
-    private var _name: String? = nil
-    private var _symbol: String? = nil
-    private var _decimals: UInt8? = nil
-    private var _hasReadProperties: Bool = false
-    
+public class ERC20: IERC20, ERC20BaseProperties {
+
+    internal var _name: String? = nil
+    internal var _symbol: String? = nil
+    internal var _decimals: UInt8? = nil
+    internal var _hasReadProperties: Bool = false
+
     public var transactionOptions: TransactionOptions
     public var web3: web3
     public var provider: Web3Provider
     public var address: EthereumAddress
-    
+
     lazy var contract: web3.web3contract = {
         let contract = self.web3.contract(Web3.Utils.erc20ABI, at: self.address, abiVersion: 2)
         precondition(contract != nil)
         return contract!
     }()
-    
+
     public init(web3: web3, provider: Web3Provider, address: EthereumAddress) {
         self.web3 = web3
         self.provider = provider
@@ -49,180 +47,192 @@ public class ERC20: IERC20 {
         mergedOptions.to = address
         self.transactionOptions = mergedOptions
     }
-    
-    public var name: String {
-        self.readProperties()
-        if self._name != nil {
-            return self._name!
-        }
-        return ""
+
+    public func getBalance(account: EthereumAddress) async throws -> BigUInt {
+        let contract = self.contract
+        var transactionOptions = TransactionOptions()
+        transactionOptions.callOnBlock = .latest
+        let result = try await contract.read("balanceOf", parameters: [account] as [AnyObject], extraData: Data(), transactionOptions: self.transactionOptions)?.call(transactionOptions: transactionOptions)
+        guard let res = result?["0"] as? BigUInt else {throw Web3Error.processingError(desc: "Failed to get result of expected type from the Ethereum node")}
+        return res
     }
-    
-    public var symbol: String {
-        self.readProperties()
-        if self._symbol != nil {
-            return self._symbol!
-        }
-        return ""
+
+    public func getAllowance(originalOwner: EthereumAddress, delegate: EthereumAddress) async throws -> BigUInt {
+        let contract = self.contract
+        var transactionOptions = TransactionOptions()
+        transactionOptions.callOnBlock = .latest
+        let result = try await contract.read("allowance", parameters: [originalOwner, delegate] as [AnyObject], extraData: Data(), transactionOptions: self.transactionOptions)?.call(transactionOptions: transactionOptions)
+        guard let res = result?["0"] as? BigUInt else {throw Web3Error.processingError(desc: "Failed to get result of expected type from the Ethereum node")}
+        return res
     }
-    
-    public var decimals: UInt8 {
-        self.readProperties()
-        if self._decimals != nil {
-            return self._decimals!
+
+    public func transfer(from: EthereumAddress, to: EthereumAddress, amount: String) async throws -> WriteTransaction {
+        let contract = self.contract
+        var basicOptions = TransactionOptions()
+        basicOptions.from = from
+        basicOptions.to = self.address
+        basicOptions.callOnBlock = .latest
+
+        // get the decimals manually
+        let callResult = try await contract.read("decimals", transactionOptions: basicOptions)?.call()
+        var decimals = BigUInt(0)
+        guard let dec = callResult?["0"], let decTyped = dec as? BigUInt else {
+                throw Web3Error.inputError(desc: "Contract may be not ERC20 compatible, can not get decimals")}
+        decimals = decTyped
+
+        let intDecimals = Int(decimals)
+        guard let value = Web3.Utils.parseToBigUInt(amount, decimals: intDecimals) else {
+            throw Web3Error.inputError(desc: "Can not parse inputted amount")
         }
-        return 255
+        guard let tx = contract.write("transfer", parameters: [to, value] as [AnyObject], transactionOptions: basicOptions) else {
+            throw Web3Error.processingError(desc: "Failed to write to contract")
+        }
+        return tx
     }
-    
-    public func readProperties() {
+
+    public func transferFrom(from: EthereumAddress, to: EthereumAddress, originalOwner: EthereumAddress, amount: String) async throws -> WriteTransaction {
+        let contract = self.contract
+        var basicOptions = TransactionOptions()
+        basicOptions.from = from
+        basicOptions.to = self.address
+        basicOptions.callOnBlock = .latest
+
+        // get the decimals manually
+        let callResult = try await contract.read("decimals", transactionOptions: basicOptions)?.call()
+        var decimals = BigUInt(0)
+        guard let dec = callResult?["0"], let decTyped = dec as? BigUInt else {
+            throw Web3Error.inputError(desc: "Contract may be not ERC20 compatible, can not get decimals")}
+        decimals = decTyped
+
+        let intDecimals = Int(decimals)
+        guard let value = Web3.Utils.parseToBigUInt(amount, decimals: intDecimals) else {
+            throw Web3Error.inputError(desc: "Can not parse inputted amount")
+        }
+
+        guard let tx = contract.write("transferFrom", parameters: [originalOwner, to, value] as [AnyObject], transactionOptions: basicOptions) else {
+            throw Web3Error.processingError(desc: "Failed to write to contract")
+        }
+        return tx
+    }
+
+    public func setAllowance(from: EthereumAddress, to: EthereumAddress, newAmount: String) async throws -> WriteTransaction {
+        let contract = self.contract
+        var basicOptions = TransactionOptions()
+        basicOptions.from = from
+        basicOptions.to = self.address
+        basicOptions.callOnBlock = .latest
+
+        // get the decimals manually
+        let callResult = try await contract.read("decimals", transactionOptions: basicOptions)?.call()
+        var decimals = BigUInt(0)
+        guard let dec = callResult?["0"], let decTyped = dec as? BigUInt else {
+            throw Web3Error.inputError(desc: "Contract may be not ERC20 compatible, can not get decimals")}
+        decimals = decTyped
+
+        let intDecimals = Int(decimals)
+        guard let value = Web3.Utils.parseToBigUInt(newAmount, decimals: intDecimals) else {
+            throw Web3Error.inputError(desc: "Can not parse inputted amount")
+        }
+
+        guard let tx = contract.write("setAllowance", parameters: [to, value] as [AnyObject], transactionOptions: basicOptions) else {
+            throw Web3Error.processingError(desc: "Failed to write to contract")
+        }
+        return tx
+    }
+
+    public func approve(from: EthereumAddress, spender: EthereumAddress, amount: String) async throws -> WriteTransaction {
+        let contract = self.contract
+        var basicOptions = TransactionOptions()
+        basicOptions.from = from
+        basicOptions.to = self.address
+        basicOptions.callOnBlock = .latest
+
+        // get the decimals manually
+        let callResult = try await contract.read("decimals", transactionOptions: basicOptions)?.call()
+        var decimals = BigUInt(0)
+        guard let dec = callResult?["0"], let decTyped = dec as? BigUInt else {
+            throw Web3Error.inputError(desc: "Contract may be not ERC20 compatible, can not get decimals")}
+        decimals = decTyped
+
+        let intDecimals = Int(decimals)
+        guard let value = Web3.Utils.parseToBigUInt(amount, decimals: intDecimals) else {
+            throw Web3Error.inputError(desc: "Can not parse inputted amount")
+        }
+
+        guard let tx = contract.write("approve", parameters: [spender, value] as [AnyObject], transactionOptions: basicOptions) else {
+            throw Web3Error.processingError(desc: "Failed to write to contract")
+        }
+        return tx
+    }
+
+    public func totalSupply() async throws -> BigUInt {
+        let contract = self.contract
+        var transactionOptions = TransactionOptions()
+        transactionOptions.callOnBlock = .latest
+        let result = try await contract.read("totalSupply", parameters: [AnyObject](), extraData: Data(), transactionOptions: self.transactionOptions)?.call(transactionOptions: transactionOptions)
+        guard let res = result?["0"] as? BigUInt else {throw Web3Error.processingError(desc: "Failed to get result of expected type from the Ethereum node")}
+        return res
+    }
+
+}
+
+protocol ERC20BaseProperties: AnyObject {
+    var contract: web3.web3contract { get }
+    var _name: String? { get set }
+    var _symbol: String? { get set }
+    var _decimals: UInt8? { get set }
+    var _hasReadProperties: Bool { get set }
+    func readProperties() async throws
+    func name() async throws -> String
+    func symbol() async throws -> String
+    func decimals() async throws -> UInt8
+}
+
+extension ERC20BaseProperties {
+    public func name() async throws -> String {
+        try await self.readProperties()
+        return self._name ?? ""
+    }
+
+    public func symbol() async throws -> String {
+        try await self.readProperties()
+        return self._symbol ?? ""
+    }
+
+    public func decimals() async throws -> UInt8 {
+        try await self.readProperties()
+        return self._decimals ?? 255
+    }
+
+    func readProperties() async throws {
         if self._hasReadProperties {
             return
         }
         let contract = self.contract
         guard contract.contract.address != nil else {return}
-        var transactionOptions = TransactionOptions.defaultOptions
-        transactionOptions.callOnBlock = .latest
-        guard let namePromise = contract.read("name", parameters: [] as [AnyObject], extraData: Data(), transactionOptions: transactionOptions)?.callPromise() else {return}
-        
-        guard let symbolPromise = contract.read("symbol", parameters: [] as [AnyObject], extraData: Data(), transactionOptions: transactionOptions)?.callPromise() else {return}
-        
-        guard let decimalPromise = contract.read("decimals", parameters: [] as [AnyObject], extraData: Data(), transactionOptions: transactionOptions)?.callPromise() else {return}
+        var transactionOptionsVAR = TransactionOptions.defaultOptions
+        transactionOptionsVAR.callOnBlock = .latest
+        let transactionOptions = transactionOptionsVAR
+        async let namePromise = contract.read("name", parameters: [AnyObject](), extraData: Data(), transactionOptions: transactionOptions)?.callPromise()
 
-        let allPromises = [namePromise, symbolPromise, decimalPromise]
-        let queue = self.web3.requestDispatcher.queue
-        when(resolved: allPromises).map(on: queue) { (resolvedPromises) -> Void in
-            guard case .fulfilled(let nameResult) = resolvedPromises[0] else {return}
-            guard let name = nameResult["0"] as? String else {return}
-            self._name = name
-            
-            guard case .fulfilled(let symbolResult) = resolvedPromises[1] else {return}
-            guard let symbol = symbolResult["0"] as? String else {return}
-            self._symbol = symbol
-            
-            guard case .fulfilled(let decimalsResult) = resolvedPromises[2] else {return}
-            guard let decimals = decimalsResult["0"] as? BigUInt else {return}
-            self._decimals = UInt8(decimals)
-            
-            self._hasReadProperties = true
-        }.wait()
-    }
-    
-    public func getBalance(account: EthereumAddress) throws -> BigUInt {
-        let contract = self.contract
-        var transactionOptions = TransactionOptions()
-        transactionOptions.callOnBlock = .latest
-        let result = try contract.read("balanceOf", parameters: [account] as [AnyObject], extraData: Data(), transactionOptions: self.transactionOptions)!.call(transactionOptions: transactionOptions)
-        guard let res = result["0"] as? BigUInt else {throw Web3Error.processingError(desc: "Failed to get result of expected type from the Ethereum node")}
-        return res
-    }
-    
-    public func getAllowance(originalOwner: EthereumAddress, delegate: EthereumAddress) throws -> BigUInt {
-        let contract = self.contract
-        var transactionOptions = TransactionOptions()
-        transactionOptions.callOnBlock = .latest
-        let result = try contract.read("allowance", parameters: [originalOwner, delegate] as [AnyObject], extraData: Data(), transactionOptions: self.transactionOptions)!.call(transactionOptions: transactionOptions)
-        guard let res = result["0"] as? BigUInt else {throw Web3Error.processingError(desc: "Failed to get result of expected type from the Ethereum node")}
-        return res
-    }
-    
-    public func transfer(from: EthereumAddress, to: EthereumAddress, amount: String) throws -> WriteTransaction {
-        let contract = self.contract
-        var basicOptions = TransactionOptions()
-        basicOptions.from = from
-        basicOptions.to = self.address
-        basicOptions.callOnBlock = .latest
-        
-        // get the decimals manually
-        let callResult = try contract.read("decimals", transactionOptions: basicOptions)!.call()
-        var decimals = BigUInt(0)
-        guard let dec = callResult["0"], let decTyped = dec as? BigUInt else {
-                throw Web3Error.inputError(desc: "Contract may be not ERC20 compatible, can not get decimals")}
-        decimals = decTyped
-        
-        let intDecimals = Int(decimals)
-        guard let value = Web3.Utils.parseToBigUInt(amount, decimals: intDecimals) else {
-            throw Web3Error.inputError(desc: "Can not parse inputted amount")
-        }
-        let tx = contract.write("transfer", parameters: [to, value] as [AnyObject], transactionOptions: basicOptions)!
-        return tx
-    }
+        async let symbolPromise = try await contract.read("symbol", parameters: [AnyObject](), extraData: Data(), transactionOptions: transactionOptions)?.callPromise()
 
-    public func transferFrom(from: EthereumAddress, to: EthereumAddress, originalOwner: EthereumAddress, amount: String) throws -> WriteTransaction {
-        let contract = self.contract
-        var basicOptions = TransactionOptions()
-        basicOptions.from = from
-        basicOptions.to = self.address
-        basicOptions.callOnBlock = .latest
-        
-        // get the decimals manually
-        let callResult = try contract.read("decimals", transactionOptions: basicOptions)!.call()
-        var decimals = BigUInt(0)
-        guard let dec = callResult["0"], let decTyped = dec as? BigUInt else {
-            throw Web3Error.inputError(desc: "Contract may be not ERC20 compatible, can not get decimals")}
-        decimals = decTyped
-        
-        let intDecimals = Int(decimals)
-        guard let value = Web3.Utils.parseToBigUInt(amount, decimals: intDecimals) else {
-            throw Web3Error.inputError(desc: "Can not parse inputted amount")
-        }
-        
-        let tx = contract.write("transferFrom", parameters: [originalOwner, to, value] as [AnyObject], transactionOptions: basicOptions)!
-        return tx
-    }
+        async let decimalPromise = try await contract.read("decimals", parameters: [AnyObject](), extraData: Data(), transactionOptions: transactionOptions)?.callPromise()
 
-    public func setAllowance(from: EthereumAddress, to: EthereumAddress, newAmount: String) throws -> WriteTransaction {
-        let contract = self.contract
-        var basicOptions = TransactionOptions()
-        basicOptions.from = from
-        basicOptions.to = self.address
-        basicOptions.callOnBlock = .latest
-        
-        // get the decimals manually
-        let callResult = try contract.read("decimals", transactionOptions: basicOptions)!.call()
-        var decimals = BigUInt(0)
-        guard let dec = callResult["0"], let decTyped = dec as? BigUInt else {
-            throw Web3Error.inputError(desc: "Contract may be not ERC20 compatible, can not get decimals")}
-        decimals = decTyped
-        
-        let intDecimals = Int(decimals)
-        guard let value = Web3.Utils.parseToBigUInt(newAmount, decimals: intDecimals) else {
-            throw Web3Error.inputError(desc: "Can not parse inputted amount")
+        let resolvedPromises = try await ["name":namePromise, "symbol":symbolPromise, "decimals":decimalPromise]
+
+        if let nameResult = resolvedPromises["name"], let name = nameResult?["0"] as? String {
+            _name = name
         }
-        
-        let tx = contract.write("setAllowance", parameters: [to, value] as [AnyObject], transactionOptions: basicOptions)!
-        return tx
-    }
-    
-    public func approve(from: EthereumAddress, spender: EthereumAddress, amount: String) throws -> WriteTransaction {
-        let contract = self.contract
-        var basicOptions = TransactionOptions()
-        basicOptions.from = from
-        basicOptions.to = self.address
-        basicOptions.callOnBlock = .latest
-        
-        // get the decimals manually
-        let callResult = try contract.read("decimals", transactionOptions: basicOptions)!.call()
-        var decimals = BigUInt(0)
-        guard let dec = callResult["0"], let decTyped = dec as? BigUInt else {
-            throw Web3Error.inputError(desc: "Contract may be not ERC20 compatible, can not get decimals")}
-        decimals = decTyped
-        
-        let intDecimals = Int(decimals)
-        guard let value = Web3.Utils.parseToBigUInt(amount, decimals: intDecimals) else {
-            throw Web3Error.inputError(desc: "Can not parse inputted amount")
+
+        if let symbolResult = resolvedPromises["symbol"], let symbol = symbolResult?["0"] as? String {
+            _symbol = symbol
         }
-        
-        let tx = contract.write("approve", parameters: [spender, value] as [AnyObject], transactionOptions: basicOptions)!
-        return tx
+
+        if let decimalsResult = resolvedPromises["decimals"], let decimals = decimalsResult?["0"] as? BigUInt {
+            _decimals = UInt8(decimals)
+        }
+
+        self._hasReadProperties = true
     }
-    
-    public func totalSupply() throws -> BigUInt {
-        let contract = self.contract
-        var transactionOptions = TransactionOptions()
-        transactionOptions.callOnBlock = .latest
-        let result = try contract.read("totalSupply", parameters: [AnyObject](), extraData: Data(), transactionOptions: self.transactionOptions)!.call(transactionOptions: transactionOptions)
-        guard let res = result["0"] as? BigUInt else {throw Web3Error.processingError(desc: "Failed to get result of expected type from the Ethereum node")}
-        return res
-    }
-    
 }
