@@ -10,38 +10,38 @@ import XCTest
 
 // MARK: Works only with network connection
 class web3swift_infura_Tests: XCTestCase {
-    
-    func testGetBalance() throws {
+
+    func testGetBalance() async {
         do {
-            let web3 = Web3.InfuraMainnetWeb3()
+            let web3 = try Web3.InfuraMainnetWeb3()
             let address = EthereumAddress("0xe22b8979739D724343bd002F9f432F5990879901")!
-            let balance = try web3.eth.getBalance(address: address)
+            let balance = try await web3.eth.getBalance(address: address)
             let balString = Web3.Utils.formatToEthereumUnits(balance, toUnits: .eth, decimals: 3)
             print(balString!)
         } catch {
             XCTFail()
         }
     }
-    
-    func testGetBlockByHash() throws {
+
+    func testGetBlockByHash() async throws {
         do {
-            let web3 = Web3.InfuraMainnetWeb3()
-            let result = try web3.eth.getBlockByHash("0x6d05ba24da6b7a1af22dc6cc2a1fe42f58b2a5ea4c406b19c8cf672ed8ec0695", fullTransactions: true)
+            let web3 = try Web3.InfuraMainnetWeb3()
+            let result = try await web3.eth.getBlockByHash("0x6d05ba24da6b7a1af22dc6cc2a1fe42f58b2a5ea4c406b19c8cf672ed8ec0695", fullTransactions: true)
             print(result)
         } catch {
             XCTFail()
         }
     }
-    
-    func testGetBlockByNumber1() throws {
-        let web3 = Web3.InfuraMainnetWeb3()
-        let result = try web3.eth.getBlockByNumber("latest", fullTransactions: true)
+
+    func testGetBlockByNumber1() async throws {
+        let web3 = try Web3.InfuraMainnetWeb3()
+        let result = try await web3.eth.getBlockByNumber("latest", fullTransactions: true)
         print(result)
     }
-    
-    func testGetBlockByNumber2() throws {
-        let web3 = Web3.InfuraMainnetWeb3()
-        let result = try web3.eth.getBlockByNumber(UInt64(5184323), fullTransactions: true)
+
+    func testGetBlockByNumber2() async throws {
+        let web3 = try Web3.InfuraMainnetWeb3()
+        let result = try await web3.eth.getBlockByNumber(UInt64(5184323), fullTransactions: true)
         print(result)
         let transactions = result.transactions
         for transaction in transactions {
@@ -53,59 +53,59 @@ class web3swift_infura_Tests: XCTestCase {
             }
         }
     }
-    
-    func testGetBlockByNumber3() throws {
+
+    func testGetBlockByNumber3() async throws {
         do {
-            let web3 = Web3.InfuraMainnetWeb3()
-            let _ = try web3.eth.getBlockByNumber(UInt64(1000000000), fullTransactions: true)
+            let web3 = try Web3.InfuraMainnetWeb3()
+            let _ = try await web3.eth.getBlockByNumber(UInt64(1000000000), fullTransactions: true)
             XCTFail()
         } catch {
-            
+
         }
     }
-    
-    func testGasPrice() throws {
-        let web3 = Web3.InfuraMainnetWeb3()
-        let response = try web3.eth.getGasPrice()
+
+    func testGasPrice() async throws {
+        let web3 = try Web3.InfuraMainnetWeb3()
+        let response = try await web3.eth.getGasPrice()
         print(response)
     }
-    
-    func testGetIndexedEventsPromise() throws {
+
+    func testGetIndexedEventsPromise() async throws {
         let jsonString = "[{\"constant\":true,\"inputs\":[],\"name\":\"name\",\"outputs\":[{\"name\":\"\",\"type\":\"string\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":false,\"inputs\":[{\"name\":\"_spender\",\"type\":\"address\"},{\"name\":\"_value\",\"type\":\"uint256\"}],\"name\":\"approve\",\"outputs\":[{\"name\":\"success\",\"type\":\"bool\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":true,\"inputs\":[],\"name\":\"totalSupply\",\"outputs\":[{\"name\":\"\",\"type\":\"uint256\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":false,\"inputs\":[{\"name\":\"_from\",\"type\":\"address\"},{\"name\":\"_to\",\"type\":\"address\"},{\"name\":\"_value\",\"type\":\"uint256\"}],\"name\":\"transferFrom\",\"outputs\":[{\"name\":\"success\",\"type\":\"bool\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":true,\"inputs\":[],\"name\":\"decimals\",\"outputs\":[{\"name\":\"\",\"type\":\"uint8\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":true,\"inputs\":[],\"name\":\"version\",\"outputs\":[{\"name\":\"\",\"type\":\"string\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":true,\"inputs\":[{\"name\":\"_owner\",\"type\":\"address\"}],\"name\":\"balanceOf\",\"outputs\":[{\"name\":\"balance\",\"type\":\"uint256\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":true,\"inputs\":[],\"name\":\"symbol\",\"outputs\":[{\"name\":\"\",\"type\":\"string\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":false,\"inputs\":[{\"name\":\"_to\",\"type\":\"address\"},{\"name\":\"_value\",\"type\":\"uint256\"}],\"name\":\"transfer\",\"outputs\":[{\"name\":\"success\",\"type\":\"bool\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":false,\"inputs\":[{\"name\":\"_spender\",\"type\":\"address\"},{\"name\":\"_value\",\"type\":\"uint256\"},{\"name\":\"_extraData\",\"type\":\"bytes\"}],\"name\":\"approveAndCall\",\"outputs\":[{\"name\":\"success\",\"type\":\"bool\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":true,\"inputs\":[{\"name\":\"_owner\",\"type\":\"address\"},{\"name\":\"_spender\",\"type\":\"address\"}],\"name\":\"allowance\",\"outputs\":[{\"name\":\"remaining\",\"type\":\"uint256\"}],\"payable\":false,\"type\":\"function\"},{\"inputs\":[{\"name\":\"_initialAmount\",\"type\":\"uint256\"},{\"name\":\"_tokenName\",\"type\":\"string\"},{\"name\":\"_decimalUnits\",\"type\":\"uint8\"},{\"name\":\"_tokenSymbol\",\"type\":\"string\"}],\"type\":\"constructor\"},{\"payable\":false,\"type\":\"fallback\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":true,\"name\":\"_from\",\"type\":\"address\"},{\"indexed\":true,\"name\":\"_to\",\"type\":\"address\"},{\"indexed\":false,\"name\":\"_value\",\"type\":\"uint256\"}],\"name\":\"Transfer\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":true,\"name\":\"_owner\",\"type\":\"address\"},{\"indexed\":true,\"name\":\"_spender\",\"type\":\"address\"},{\"indexed\":false,\"name\":\"_value\",\"type\":\"uint256\"}],\"name\":\"Approval\",\"type\":\"event\"},]"
-        let web3 = Web3.InfuraMainnetWeb3()
+        let web3 = try Web3.InfuraMainnetWeb3()
         let contract = web3.contract(jsonString, at: nil, abiVersion: 2)
         var filter = EventFilter()
         filter.fromBlock = .blockNumber(UInt64(5200120))
         filter.toBlock = .blockNumber(UInt64(5200120))
         filter.addresses = [EthereumAddress("0x53066cddbc0099eb6c96785d9b3df2aaeede5da3")!]
         filter.parameterFilters = [([EthereumAddress("0xefdcf2c36f3756ce7247628afdb632fa4ee12ec5")!] as [EventFilterable]), (nil as [EventFilterable]?)]
-        let eventParserResult = try contract!.getIndexedEventsPromise(eventName: "Transfer", filter: filter, joinWithReceipts: true).wait()
+        let eventParserResult = try await contract!.getIndexedEventsPromise(eventName: "Transfer", filter: filter, joinWithReceipts: true)
         print(eventParserResult)
         XCTAssert(eventParserResult.count == 2)
         XCTAssert(eventParserResult[0].transactionReceipt != nil)
         XCTAssert(eventParserResult[0].eventLog != nil)
     }
-    
-    func testEventParsingBlockByNumberPromise() throws {
+
+    func testEventParsingBlockByNumberPromise() async throws {
         let jsonString = "[{\"constant\":true,\"inputs\":[],\"name\":\"name\",\"outputs\":[{\"name\":\"\",\"type\":\"string\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":false,\"inputs\":[{\"name\":\"_spender\",\"type\":\"address\"},{\"name\":\"_value\",\"type\":\"uint256\"}],\"name\":\"approve\",\"outputs\":[{\"name\":\"success\",\"type\":\"bool\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":true,\"inputs\":[],\"name\":\"totalSupply\",\"outputs\":[{\"name\":\"\",\"type\":\"uint256\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":false,\"inputs\":[{\"name\":\"_from\",\"type\":\"address\"},{\"name\":\"_to\",\"type\":\"address\"},{\"name\":\"_value\",\"type\":\"uint256\"}],\"name\":\"transferFrom\",\"outputs\":[{\"name\":\"success\",\"type\":\"bool\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":true,\"inputs\":[],\"name\":\"decimals\",\"outputs\":[{\"name\":\"\",\"type\":\"uint8\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":true,\"inputs\":[],\"name\":\"version\",\"outputs\":[{\"name\":\"\",\"type\":\"string\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":true,\"inputs\":[{\"name\":\"_owner\",\"type\":\"address\"}],\"name\":\"balanceOf\",\"outputs\":[{\"name\":\"balance\",\"type\":\"uint256\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":true,\"inputs\":[],\"name\":\"symbol\",\"outputs\":[{\"name\":\"\",\"type\":\"string\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":false,\"inputs\":[{\"name\":\"_to\",\"type\":\"address\"},{\"name\":\"_value\",\"type\":\"uint256\"}],\"name\":\"transfer\",\"outputs\":[{\"name\":\"success\",\"type\":\"bool\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":false,\"inputs\":[{\"name\":\"_spender\",\"type\":\"address\"},{\"name\":\"_value\",\"type\":\"uint256\"},{\"name\":\"_extraData\",\"type\":\"bytes\"}],\"name\":\"approveAndCall\",\"outputs\":[{\"name\":\"success\",\"type\":\"bool\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":true,\"inputs\":[{\"name\":\"_owner\",\"type\":\"address\"},{\"name\":\"_spender\",\"type\":\"address\"}],\"name\":\"allowance\",\"outputs\":[{\"name\":\"remaining\",\"type\":\"uint256\"}],\"payable\":false,\"type\":\"function\"},{\"inputs\":[{\"name\":\"_initialAmount\",\"type\":\"uint256\"},{\"name\":\"_tokenName\",\"type\":\"string\"},{\"name\":\"_decimalUnits\",\"type\":\"uint8\"},{\"name\":\"_tokenSymbol\",\"type\":\"string\"}],\"type\":\"constructor\"},{\"payable\":false,\"type\":\"fallback\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":true,\"name\":\"_from\",\"type\":\"address\"},{\"indexed\":true,\"name\":\"_to\",\"type\":\"address\"},{\"indexed\":false,\"name\":\"_value\",\"type\":\"uint256\"}],\"name\":\"Transfer\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":true,\"name\":\"_owner\",\"type\":\"address\"},{\"indexed\":true,\"name\":\"_spender\",\"type\":\"address\"},{\"indexed\":false,\"name\":\"_value\",\"type\":\"uint256\"}],\"name\":\"Approval\",\"type\":\"event\"},]"
-        let web3 = Web3.InfuraMainnetWeb3()
+        let web3 = try Web3.InfuraMainnetWeb3()
         let contract = web3.contract(jsonString, at: nil, abiVersion: 2)
         var filter = EventFilter()
         filter.addresses = [EthereumAddress("0x53066cddbc0099eb6c96785d9b3df2aaeede5da3")!]
         filter.parameterFilters = [([EthereumAddress("0xefdcf2c36f3756ce7247628afdb632fa4ee12ec5")!] as [EventFilterable]), ([EthereumAddress("0xd5395c132c791a7f46fa8fc27f0ab6bacd824484")!] as [EventFilterable])]
-        guard let eventParser = contract?.createEventParser("Transfer", filter: filter) else {return XCTFail()}
-        let present = try eventParser.parseBlockByNumberPromise(UInt64(5200120)).wait()
+        guard let eventParser = try contract?.createEventParser("Transfer", filter: filter) else {return XCTFail()}
+        let present = try await eventParser.parseBlockByNumberPromise(UInt64(5200120))
         print(present)
         XCTAssert(present.count == 1)
     }
-    
-    func testUserCaseEventParsing() throws {
+
+    func testUserCaseEventParsing() async throws {
         let contractAddress = EthereumAddress("0x7ff546aaccd379d2d1f241e1d29cdd61d4d50778")
         let jsonString = "[{\"constant\":false,\"inputs\":[{\"name\":\"_id\",\"type\":\"string\"}],\"name\":\"deposit\",\"outputs\":[],\"payable\":true,\"stateMutability\":\"payable\",\"type\":\"function\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":true,\"name\":\"_from\",\"type\":\"address\"},{\"indexed\":false,\"name\":\"_id\",\"type\":\"string\"},{\"indexed\":true,\"name\":\"_value\",\"type\":\"uint256\"}],\"name\":\"Deposit\",\"type\":\"event\"}]"
-        let web3 = Web3.InfuraRinkebyWeb3()
+        let web3 = try Web3.InfuraRinkebyWeb3()
         let contract = web3.contract(jsonString, at: contractAddress, abiVersion: 2)
-        guard let eventParser = contract?.createEventParser("Deposit", filter: nil) else {return XCTFail()}
-        let pres = try eventParser.parseBlockByNumber(UInt64(2138657))
+        guard let eventParser = try contract?.createEventParser("Deposit", filter: nil) else {return XCTFail()}
+        let pres = try await eventParser.parseBlockByNumber(UInt64(2138657))
         XCTAssert(pres.count == 1)
     }
 }
