@@ -10,29 +10,29 @@ import BigInt
 @testable import web3swift
 
 class web3swift_ERC20_Tests: XCTestCase {
-    
-    func testERC20name() throws {
-        let (web3, _, receipt, _) = try web3swift_helpers.localDeployERC20()
-        
+
+    func testERC20name() async throws {
+        let (web3, _, receipt, _) = try await web3swift_helpers.localDeployERC20()
+
         let parameters = [] as [AnyObject]
         let contract = web3.contract(Web3.Utils.erc20ABI, at: receipt.contractAddress!)!
         let readTX = contract.read("name", parameters:parameters)!
         readTX.transactionOptions.from = EthereumAddress("0xe22b8979739D724343bd002F9f432F5990879901")
-        let response = try readTX.callPromise().wait()
+        let response = try await readTX.callPromise()
         let name = response["0"] as? String
         XCTAssert(name == "web3swift", "Failed to create ERC20 name transaction")
     }
-    
-    func testERC20tokenBalance() throws {
-        let (web3, _, receipt, _) = try web3swift_helpers.localDeployERC20()
-        
+
+    func testERC20tokenBalance() async throws {
+        let (web3, _, receipt, _) = try await web3swift_helpers.localDeployERC20()
+
         let addressOfUser = EthereumAddress("0xe22b8979739D724343bd002F9f432F5990879901")!
         let contract = web3.contract(Web3.Utils.erc20ABI, at: receipt.contractAddress!, abiVersion: 2)!
         guard let readTX = contract.read("balanceOf", parameters: [addressOfUser] as [AnyObject]) else {return XCTFail()}
         readTX.transactionOptions.from = addressOfUser
-        let tokenBalance = try readTX.callPromise().wait()
+        let tokenBalance = try await readTX.callPromise()
         guard let bal = tokenBalance["0"] as? BigUInt else {return XCTFail()}
         print(String(bal))
     }
-    
+
 }
